requirements.txt
streamlit
openai
sentry-sdk
galileo-observe
python-dotenv
daytona
